make fps a float

+++ toascii/Live.py
@@ -8,7 +8,7 @@ from .Constants import *
 
 
 class Live(ABC):
-    def __init__(self, source: int = 0, *, scale: float = 1, w_stretch: float = 2, gradient: typing.Union[int, str] = 0, fps: int = 10, verbose: int = False):
+    def __init__(self, source: int = 0, *, scale: float = 1, w_stretch: float = 2, gradient: typing.Union[int, str] = 0, fps: float = 10, verbose: int = False):
         self.source = source
         self.video = cv2.VideoCapture(self.source)
